+++ extension.php
@@ -11,7 +11,7 @@
  * bundled with this package in the LICENSE file.
  *
  * @package    Platform Pages extension
- * @version    3.0.0
+ * @version    3.1.0
  * @author     Cartalyst LLC
  * @license    Cartalyst PSL
  * @copyright  (c) 2011-2015, Cartalyst LLC
@@ -83,7 +83,7 @@ return [
     |
     */
 
-    'version' => '3.0.0',
+    'version' => '3.1.0',
 
     /*
     |--------------------------------------------------------------------------